+++ src/website/root/about/dashboard.php
@@ -15,7 +15,7 @@ $text_info = $about_obj->get_all_texts();
   </div>
   <div class="table-responsive-sm">
     <!-- strst pieces table -->
-    <table class="table table-bordered table-striped table-striped display compact table-style" style="width:100%">
+    <table class="table table-bordered table-striped table-striped display display-big-data compact table-style" style="width:100%">
       <thead class="primary text-capitalize">
         <tr>
           <th class="text-center" style="max-width: 40px">#</th>
@@ -73,22 +73,18 @@ $text_info = $about_obj->get_all_texts();
               </td>
               <td class="text-center">
                 <?php if ($text['is_active']) { ?>
-                  <a href="?do=deactivate-text&id=<?php echo base64_encode($text['id']) ?>" class="btn btn-danger py-0"><i
-                      class="bi bi-x"></i>
+                  <a href="?do=deactivate-text&id=<?php echo base64_encode($text['id']) ?>" class="btn btn-danger py-0"><i class="bi bi-x"></i>
                     <span>
                       <?php echo lang('DEACTIVATE') ?>
                     </span></a>
                 <?php } else { ?>
-                  <a href="?do=activate-text&id=<?php echo base64_encode($text['id']) ?>" class="btn btn-primary py-0"><i
-                      class="bi bi-check"></i>
+                  <a href="?do=activate-text&id=<?php echo base64_encode($text['id']) ?>" class="btn btn-primary py-0"><i class="bi bi-check"></i>
                     <span>
                       <?php echo lang('ACTIVATE') ?>
                     </span></a>
                 <?php } ?>
-                <a href="?do=edit-text&id=<?php echo base64_encode($text['id']) ?>" class="btn btn-outline-success py-0"><i
-                    class="bi bi-pencil-square"></i></a>
-                <button type="button" data-href="?do=delete-text&id=<?php echo base64_encode($text['id']) ?>&back=true"
-                  class="btn btn-outline-danger py-0" onclick="confirm_delete(this)">
+                <a href="?do=edit-text&id=<?php echo base64_encode($text['id']) ?>" class="btn btn-outline-success py-0"><i class="bi bi-pencil-square"></i></a>
+                <button type="button" data-href="?do=delete-text&id=<?php echo base64_encode($text['id']) ?>&back=true" class="btn btn-outline-danger py-0" onclick="confirm_delete(this)">
                   <i class="bi bi-trash"></i></button>
               </td>
             </tr>
