+++ src/website/root/features/dashboard.php
@@ -14,7 +14,7 @@ $features_info = $features_obj->get_all_features();
     </a>
   </div>
   <!-- strst pieces table -->
-  <table class="table table-bordered table-striped table-striped display compact table-style" style="width:100%">
+  <table class="table table-bordered table-striped table-striped display display-big-data compact table-style" style="width:100%">
     <thead class="primary text-capitalize">
       <tr>
         <th class="text-center" style="max-width: 40px">#</th>
@@ -67,21 +67,16 @@ $features_info = $features_obj->get_all_features();
             </td>
             <td class="text-center">
               <?php if ($feature['is_active']) { ?>
-                <a href="?do=deactivate-feature&id=<?php echo base64_encode($feature['id']) ?>" class="btn btn-danger py-0"><i
-                    class="bi bi-x"></i><span>
-                      <?php echo lang('DEACTIVATE') ?>
-                    </span></a>
+                <a href="?do=deactivate-feature&id=<?php echo base64_encode($feature['id']) ?>" class="btn btn-danger py-0"><i class="bi bi-x"></i><span>
+                    <?php echo lang('DEACTIVATE') ?>
+                  </span></a>
               <?php } else { ?>
-                <a href="?do=activate-feature&id=<?php echo base64_encode($feature['id']) ?>" class="btn btn-primary py-0"><i
-                    class="bi bi-check"></i><span>
-                      <?php echo lang('ACTIVATE') ?>
-                    </span></a>
+                <a href="?do=activate-feature&id=<?php echo base64_encode($feature['id']) ?>" class="btn btn-primary py-0"><i class="bi bi-check"></i><span>
+                    <?php echo lang('ACTIVATE') ?>
+                  </span></a>
               <?php } ?>
-              <a href="?do=edit-feature&id=<?php echo base64_encode($feature['id']) ?>"
-                class="btn btn-outline-success py-0"><i class="bi bi-pencil-square"></i></a>
-              <button type="button" data-href="?do=delete-feature&id=<?php echo base64_encode($feature['id']) ?>&back=true"
-                class="btn btn-outline-danger py-0" onclick="confirm_delete(this)"><i
-                  class="bi bi-trash"></i></button>
+              <a href="?do=edit-feature&id=<?php echo base64_encode($feature['id']) ?>" class="btn btn-outline-success py-0"><i class="bi bi-pencil-square"></i></a>
+              <button type="button" data-href="?do=delete-feature&id=<?php echo base64_encode($feature['id']) ?>&back=true" class="btn btn-outline-danger py-0" onclick="confirm_delete(this)"><i class="bi bi-trash"></i></button>
             </td>
           </tr>
         <?php } ?>
